requirements.txt
setuptools~=60.2.0
pandas~=2.0.1
numpy~=1.24.3
python-dateutil~=2.8.2
pycap~=2.4.0